% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metallicart.R
\name{metalli_palette}
\alias{metalli_palette}
\title{Color Palettes based on Metallica album covers}
\usage{
metalli_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of palette. Select one:
\code{kill}, \code{lightning}, \code{puppets}, \code{justice},
\code{metallica}, \code{load}, \code{reload}, \code{anger},
\code{magnetic}, \code{hardwired}, \code{combo}}

\item{n}{Number of colors desired.

Palettes contain 7 colors, except the \code{combo} palette, which has 10.}

\item{type}{Either continuous or discrete.}
}
\value{
A vector of colors.
}
\description{
R package that contains color palettes based on colours on Metallica studio album covers.
}
\details{
Yep, even the black album.

This package is based on the nycpalettes package: https://github.com/kellycotton/nycpalettes
}
\examples{
metalli_palette("anger")

}
