% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metallicart.R
\name{metalli_palette}
\alias{metalli_palette}
\title{Color Palettes based on Metallica album covers}
\usage{
metalli_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of palette. Select one:
\code{kill}, \code{kill10}, \code{lightning}, \code{lightning10},
\code{puppets}, \code{puppets10},\code{justice}, \code{justice10},
\code{metallica}, \code{metallica10}, \code{load},\code{load10},
\code{reload}, \code{reload10}, \code{anger}, \code{anger10},
\code{magnetic}, \code{magnetic10}, \code{hardwired}, \code{hardwired10},
\code{combo}}

\item{n}{Number of colors desired.

Some palettes contain 7 colors which were picked 'by hand'
The \code{combo} palette and those ending with '10' have 10 colours.
Apart from \code{combo} palette, these were produced with the
aid of the \code{colorfindr} package :
https://CRAN.R-project.org/package=colorfindr}

\item{type}{Either continuous or discrete.}
}
\value{
A vector of colors.
}
\description{
R package that contains color palettes based on colours on Metallica studio album covers.
}
\details{
Yep, even the black album.

This package is based on the nycpalettes package: https://github.com/kellycotton/nycpalettes
}
\examples{
metalli_palette("anger")

}
